extract onDemandCompile method
